// types/ethers-contracts/swap/Swap/MassDXSwap.ts
export declare namespace MassDXSwap {
  export type SwapStruct = {
    token: PromiseOrValue<string>;
    isEther: PromiseOrValue<boolean>;
    amount: PromiseOrValue<BigNumberish>;
  };

  export type SwapStructOutput = [string, boolean, BigNumber] & {
    token: string;
    isEther: boolean;
    amount: BigNumber;
  };
}